update matsim dependency and run script
