+++ pom.xml
@@ -6,7 +6,7 @@
 		<artifactId>matsim-all</artifactId>
 
 		<!-- PR-labelled release -->
-		<version>16.0-PR3085</version>
+		<version>16.0-PR3108</version>
 
 		<!-- snapshot == not recommended: rather use PR-labelled release!-->
 <!--		<version>16.0-SNAPSHOT</version>-->
@@ -147,6 +147,11 @@
 		<dependency>
 			<groupId>org.matsim.contrib</groupId>
 			<artifactId>roadpricing</artifactId>
+			<version>${matsim.version}</version>
+		</dependency>
+		<dependency>
+			<groupId>org.matsim.contrib</groupId>
+			<artifactId>small-scale-traffic-generation</artifactId>
 			<version>${matsim.version}</version>
 		</dependency>
