--- src/main/java/org/matsim/run/RunOpenBerlinScenarioGUI.java
@@ -1,38 +1,0 @@
-/* *********************************************************************** *
- * project: org.matsim.*
- *                                                                         *
- * *********************************************************************** *
- *                                                                         *
- * copyright       : (C) 2008 by the members listed in the COPYING,        *
- *                   LICENSE and WARRANTY file.                            *
- * email           : info at matsim dot org                                *
- *                                                                         *
- * *********************************************************************** *
- *                                                                         *
- *   This program is free software; you can redistribute it and/or modify  *
- *   it under the terms of the GNU General Public License as published by  *
- *   the Free Software Foundation; either version 2 of the License, or     *
- *   (at your option) any later version.                                   *
- *   See also COPYING, LICENSE and WARRANTY file                           *
- *                                                                         *
- * *********************************************************************** */
-
-package org.matsim.run;
-
-import org.matsim.run.gui.Gui;
-
-/**
- * Runs the MATSim GUI for the OpenBerlin scenario.
- */
-public final class RunOpenBerlinScenarioGUI {
-
-	private RunOpenBerlinScenarioGUI() {
-	}
-
-	public static void main(String[] args) {
-		Gui.show("MATSim GUI for matsim-berlin", RunOpenBerlinScenario.class);
-	}
-
-	//	The jar file is generated by "mvn package".  Note that the pom.xml refers at some point to this class here.
-
-}
